04_use_cases/document_processor/genai_rag_platform/requirements.txt
openai>=1.3.0
faiss-cpu
langchain>=0.1.0
python-dotenv
